Improved on generate_syscallmeta.py, now also emits detailed syscall parameter info. Fixed a bug in the ring0 syscall prehook patch, now adjusting the pointers works finally, and the hello-world testapp completes successfully. Woo!

+++ sources/include/dangless/common.h
@@ -134,6 +134,6 @@ void _assert_fail(void);
       static_assert_fail(); \
   }*/
 
-#define ARRAY_LENGTH(ARR) (sizeof(ARR) / sizeof(ARR[0]))
+#define ARRAY_LENGTH(ARR) ((size_t)(sizeof(ARR) / sizeof(ARR[0])))
 
 #endif
